requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
supabase==2.0.2
langchain==0.1.0
langchain-openai==0.0.2
langchain-groq==0.0.1
langgraph==0.0.20
openai==1.3.0
PyMuPDF==1.23.0
python-docx==1.1.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.0